Moved `vue-project` into root dir; integrated ErrorPopups changes
