+++ README.md
@@ -1,0 +1,9 @@
+## How to Run
+
+Locally
+
+* `npm run dev`
+
+For Production
+
+* `npm run build`
