+++ src/components/SimulationReport/CompareReports.vue
