+++ src/components/SimulationReport/DetailedReport.vue
