+++ src/components/SimulationReport/simulation-report-compare.vue
