+++ src/plugins/index.js
