+++ src/styles/settings.scss
