+++ vite.config.mjs
